move file to analytics folder
